Remove the "pvalue filterfield" widget from panel toolbars
This was always a performance stopgap for too-dense regions, but LZ 0.13 will support automatically hiding insignificant points without manual intervention

We may add back a more impactful usage of the filter widget in the future. For now, we'll simplify the toolbar to make room for other planned features, and see if users really want this back.